+++ R/model.r
@@ -209,6 +209,60 @@ estimate.hyper <- function(
 #' Filter genes
 #'
 #' @param dl de.lorean object
+#' @param .filter Function that gakes a list of genes as input and returns
+#'     a vector of TRUE/FALSE
+#' @param number Number to sample if filter function or genes not supplied.
+#' @param genes The genes to keep.
+#' @examples
+#' \dontrun{
+#' genes <- c('NANOG', 'DAZ1', 'SOX4', 'PRDM14')
+#' filter.genes(dl, .filter = function(x) x %in% genes)
+#' filter.genes(dl, number = 4)
+#' filter.genes(dl, genes = genes)
+#' }
+#'
+#' @export filter.genes
+#'
+filter.genes <- function(dl,
+                         .filter=function(x) x %in% genes,
+                         number=NULL,
+                         genes=sample(rownames(dl$expr), number))
+{
+    .Deprecated('filter_genes', 'DeLorean')
+    filter_genes(dl, .filter, number, genes)
+}
+
+
+#' Filter cells
+#'
+#' @param dl de.lorean object
+#' @param .filter Function that gakes a list of cells as input and returns
+#'     a vector of TRUE/FALSE
+#' @param number Number to sample if filter function or cells not supplied.
+#' @param cells The cells to keep.
+#' @examples
+#' \dontrun{
+#' cells <- c('cell1', 'cell2', 'cell3', 'cell4')
+#' filter.cells(dl, .filter = function(x) x %in% cells)
+#' filter.cells(dl, number = 4)
+#' filter.cells(dl, cells = cells)
+#' }
+#'
+#' @export filter.cells
+#'
+filter.cells <- function(dl,
+                         .filter=function(x) x %in% cells,
+                         number=NULL,
+                         cells=sample(colnames(dl$expr), number))
+{
+    .Deprecated('filter_cells', 'DeLorean')
+    filter_cells(dl, .filter, number, cells)
+}
+
+
+#' Filter genes
+#'
+#' @param dl de.lorean object
 #' @param number Number to sample if filter function or genes not supplied.
 #' @param genes The genes to keep.
 #' @param .filter Function that gakes a list of genes as input and returns
@@ -216,7 +270,7 @@ estimate.hyper <- function(
 #'
 #' @export
 #'
-filter.genes <- function(dl,
+filter_genes <- function(dl,
                          .filter = function(x) x %in% genes,
                          number = NULL,
                          genes = sample(rownames(dl$expr), number))
@@ -238,7 +292,7 @@ filter.genes <- function(dl,
 #'
 #' @export
 #'
-filter.cells <- function(dl,
+filter_cells <- function(dl,
                          .filter = function(x) x %in% cells,
                          number = NULL,
                          cells = sample(colnames(dl$expr), number))
@@ -268,7 +322,7 @@ sample.per.capture <- function(dl, cells.per.capture) {
         # %>% do(sample_n(., min(cells.per.capture, length(cell))))
         %>% do(sample.at.most(., cells.per.capture))
     )
-    filter.cells(dl, cells = sampled.cells$cell)
+    filter_cells(dl, cells = sampled.cells$cell)
 }
 
 
@@ -483,9 +537,290 @@ compile.model <- function(dl) {
       fit = compiled,
       data = stan.data,
       init = list(init.from.prior(dl)),
+      # init = make.chain.init.fn(dl),
       warmup = 1,
       iter = 1,
       chains = 1)
+  })
+}
+
+
+# Define a function to initialise the chains
+#
+# @param dl de.lorean object
+#
+make.chain.init.fn <- function(dl) {
+  function() {
+    with(dl$stan.data, {
+      # message("Creating initialisation")
+      init <- list(
+        S=dl$cell.map$S.hat,
+        tau=rnorm(C, mean=time, sd=sigma_tau),
+        psi=rlnorm(G, meanlog=mu_psi, sdlog=sigma_psi),
+        omega=rlnorm(G, meanlog=mu_omega, sdlog=sigma_omega)
+      )
+      init
+    })
+  }
+}
+
+
+even.tau.spread <- function(dl) {
+    with(dl$stan.data,
+         seq(min(time) - sigma_tau,
+             max(time) + sigma_tau,
+             length=C))
+}
+
+# From http://stackoverflow.com/questions/11094822/numbers-in-geometric-progression
+geom.series <- function(base, max) {
+  base^(0:floor(log(max, base)))
+}
+
+#' Use seriation package to find good orderings
+#'
+#' @param dl de.lorean object
+#' @param .methods The seriation methods to apply
+#' @param scaled Whether to use the scaled and/or unscaled expression data
+#' @param dim.red Dimension reduction methods to apply
+#' @param dims Number of dimensions to reduce to
+#' @param num.cores Number of cores to use in parallel
+#' @param num.tau.to.keep How many initialisations to keep
+#'
+#' @export
+#'
+seriation.find.orderings <- function(
+  dl,
+  # .methods = c("ARSA", "TSP", "R2E", "HC", "GW", "OLO"),
+  .methods = c("TSP", "R2E", "HC", "GW", "OLO"),
+  scaled = c('scaled', 'unscaled'),
+  dim.red = c('none', 'pca', 'kfa', 'ica', 'mds'),
+  # dim.red = c('mds'),
+  dims = geom.series(base=2, max=min(8, nrow(dl$expr)-1)),
+  num.cores = default.num.cores(),
+  num.tau.to.keep = default.num.cores())
+{
+  #
+  # Calculate all combinations of parameters
+  combinations <- expand.grid(method=.methods,
+                              scaled=scaled,
+                              dim.red=dim.red,
+                              dims=dims,
+                              stringsAsFactors=FALSE) %>%
+    # Only keep one combination with no dimensionality reduction
+    dplyr::filter(dim.red != 'none' | 1 == dims) %>%
+    # Cannot do KFA with 1 dimension
+    dplyr::filter(dim.red != 'kfa' | 1 != dims)
+  get.expr <- function(scaled) switch(scaled,
+    scaled = scale(t(dl$expr)),
+    unscaled = t(dl$expr),
+    stop('scaled must be "scaled" or "unscaled"')
+  )
+  get.expr.mem <- memoise::memoise(get.expr)
+  get.red <- function(scaled, dim.red, dims=5) {
+    expr <- get.expr(scaled)
+    switch(dim.red,
+      none = expr,
+      pca = prcomp(expr)$x[,1:dims],
+      ica = fastICA::fastICA(expr, n.comp=dims)$S,
+      kfa = t(kernlab::kfa(t(expr), features=dims)@xmatrix),
+      mds = cmdscale(dist(expr), k=dims),
+      stop('dim.red must be "none", "ica", "kfa", "mds" or "pca"'))
+  }
+  get.red.mem <- memoise::memoise(get.red)
+  get.dist <- function(scaled, dim.red, dims=5) {
+    dist(get.red.mem(scaled, dim.red, dims))
+  }
+  get.dist.mem <- memoise::memoise(get.dist)
+  result <- parallel::mclapply(
+    1:nrow(combinations),
+    function(i) with(combinations[i,], within(list(), {
+      method.name <- stringr::str_c(method,
+                                    ':', scaled,
+                                    ':', dim.red,
+                                    ':', dims)
+      elapsed <- system.time(
+        per <- seriation::seriate(get.dist.mem(scaled, dim.red, dims),
+                                   method=method))
+      ser.order <- seriation::get_order(per)
+      ll <- dl$ordering.ll(ser.order)
+      message(method.name, '; time=', elapsed[3], 's; LL=', ll)
+    })),
+    mc.cores=num.cores)
+  result
+}
+
+
+#' Run a find good ordering method and append results to existing orderings
+#'
+#' @param dl de.lorean object
+#' @param method Function that runs the method
+#' @param ... Any other arguments for the method
+#'
+#' @export
+#'
+find.good.ordering <- function(dl, method, ...)
+{
+  dl <- create.ordering.ll.fn(dl)
+  dl$order.inits <- c(
+    dl$order.inits,
+    lapply(
+      method(dl, ...),
+      function(i) {
+        i$ser.order <- rev.order.if.better(dl, i$ser.order)
+        i
+      }))
+  dl
+}
+
+
+#' Plot likelihoods of orderings against elapsed times taken
+#' to generate them
+#'
+#' @param dl The DeLorean object
+#'
+#' @export
+#'
+orderings.plot <- function(dl) with(dl, {
+  results.df <- data.frame(
+    method=sapply(order.inits, function(r) r$method.name),
+    elapsed=sapply(order.inits, function(r) r$elapsed[3]),
+    ll=sapply(order.inits, function(r) r$ll))
+  ggplot2::ggplot(results.df, aes(x=elapsed, y=ll, label=method)) +
+    ggplot2::geom_text()
+})
+
+# Reverse ordering if it is better correlated with observed times
+rev.order.if.better <- function(dl, ser.order) {
+  rev.order <- rev(ser.order)
+  if (cor(ser.order, dl$cell.map$obstime) < cor(rev.order, dl$cell.map$obstime)) {
+    ser.order <- rev.order
+  }
+  ser.order
+}
+
+# Deduplicate orderings
+#
+deduplicate.orderings <- function(dl) {
+  orderings <- sapply(dl$order.inits, function(i) i$ser.order)
+  dupd <- duplicated(t(orderings))
+  dl$order.inits <- dl$order.inits[!dupd]
+  dl
+}
+
+#' Convert best orderings into initialisations
+#'
+#' @param dl The DeLorean object
+#' @param num.to.keep The number to keep (defaults to default.num.cores())
+#'
+#' @export
+#'
+pseudotimes.from.orderings <- function(
+  dl,
+  num.to.keep=default.num.cores())
+within(deduplicate.orderings(dl), {
+  order.orderings <- order(sapply(order.inits, function(i) i$ll),
+                           decreasing=TRUE)
+  #
+  # Make sure we don't try to keep too many (due to duplicates, etc...)
+  actually.keep <- min(length(order.orderings), num.to.keep)
+  if (actually.keep < num.to.keep) {
+    warning("Don't have enough ", num.to.keep, " pseudotimes to keep,",
+            " only have ", actually.keep)
+  }
+  best.orderings <- order.inits[order.orderings[1:actually.keep]]
+  tau.inits <- lapply(
+    best.orderings,
+    function(O) {
+      message('Using ordering ', O$method.name, '; LL=', O$ll)
+      # Create an initialisation using the ordering
+      init <- init.chain.sample.tau(dl)
+      init$tau <- even.tau.spread(dl)[O$ser.order]
+      init
+    })
+})
+
+
+#' Find best order of the samples assuming some smooth GP prior on the
+#' expression profiles over this ordering.
+#'
+#' @param dl de.lorean object
+#' @param psi Temporal variation
+#' @param omega Noise
+#' @param num.cores Number of cores to run on. Defaults to default.num.cores()
+#' @param num.tau.to.try How many initialisations to try
+#' @param num.tau.to.keep How many initialisations to keep
+#' @param method Method to use "maximise" or "metropolis"
+#' @param ... Extra arguments to method
+#'
+#' @export
+#'
+find.smooth.tau <- function(
+    dl,
+    psi = exp(dl$hyper$mu_psi),
+    omega = exp(dl$hyper$mu_omega),
+    num.cores = default.num.cores(),
+    num.tau.to.try = num.cores,
+    num.tau.to.keep = num.cores,
+    method = "metropolis",
+    ...
+) {
+  dl <- create.ordering.ll.fn(dl)
+  log.likelihood <- dl$ordering.ll
+  dl$tau.inits <- with(dl, {
+    # Maximise the sum of the log marginal likelihoods
+    ordering.search <- function(seed) {
+      set.seed(seed)
+      # Choose a starting point by random projection
+      expr.centre <- t(scale(t(expr), center=T, scale=F))
+      init.ordering <- order(rnorm(nrow(expr.centre)) %*% expr.centre)
+      # init.ordering <- sample(stan.data$C)
+      metropolis.fn <- function(ordering, log.likelihood, ...) {
+        mh.run <- ordering.metropolis.hastings(
+          ordering,
+          log.likelihood,
+          proposal.fn=ordering.random.block.move,
+          ...)
+        best.sample <- which.max(mh.run$log.likelihoods)
+        #ordering.maximise(mh.run$chain[best.sample,], log.likelihood)
+        mh.run$chain[best.sample,]
+      }
+      method.fn <- switch(method,
+                          "maximise"=ordering.maximise,
+                          "metropolis"=metropolis.fn,
+                          NA)
+      ordering <- method.fn(init.ordering, log.likelihood, ...)
+      stopifnot(! is.null(ordering))
+      # Reverse the ordering if it makes it correlate better with
+      # the capture times
+      capture.order <- order(stan.data$time)
+      if (cor(capture.order, ordering) <
+          cor(capture.order, rev(ordering)))
+      {
+        ordering <- rev(ordering)
+      }
+      ordering
+    }
+    # Choose seeds
+    seeds <- sample.int(.Machine$integer.max, num.tau.to.try)
+    # Run in parallel or not?
+    if (num.cores > 1) {
+      orderings <- parallel::mclapply(seeds,
+                                      mc.cores=num.cores,
+                                      ordering.search)
+    } else {
+      orderings <- lapply(seeds, ordering.search)
+    }
+    # Order the taus by the best orderings
+    lls <- sapply(orderings, function(o) -log.likelihood(o))
+    best.order <- order(lls)
+    # Make the complete chain initialisation with the tau.
+    lapply(orderings[best.order[1:num.tau.to.keep]],
+           function(ordering) {
+             init <- init.chain.sample.tau(dl)
+             init$tau <- even.tau.spread(dl)[ordering.invert(ordering)]
+             init
+           })
   })
 }
 
@@ -792,10 +1127,11 @@ process.posterior <- function(dl) {
                                    model.parameter.dimensions(dl))
     best.sample <- which.max(samples.l$lp__$lp__)
     if (TRUE %in% samples.l$logmarglike$is.held.out) {
-      mean.held.out.marg.ll <- mean(
-        (samples.l$logmarglike
-        %>% left_join(gene.map)
-        %>% filter(is.held.out))$logmarglike)
+      mean.held.out.marg.ll <-
+        mean((
+          samples.l$logmarglike %>%
+          left_join(gene.map) %>%
+          filter(is.held.out))$logmarglike)
       message('Mean held out marginal log likelihood per cell: ',
               mean.held.out.marg.ll / stan.data$C)
     }
@@ -808,7 +1144,7 @@ process.posterior <- function(dl) {
     #
     # Expose Stan functions from model
     stan.fns <- new.env()
-    rstan::expose_stan_functions(fit, stan.fns)
+    rstan::expose_stan_functions(fit, env = stan.fns)
   })
 }